Optimized the import of dependencies for Large Language Models.
